Refactor slightly #32
